feat: Add a public rectangle struct named `Window`
Add a new public type representing a rectangular window on the canvas.
It will be used as an argument for the windowed image export methods.

Add struct definition docstrings and doctests.
Add a constructor and a few utility methods.
Define some common traits for the new type.

TODO: Add windowed image export methods to `Canvas`.